requirements.txt
Flask==3.0.3
Flask-APScheduler==1.13.1
line-bot-sdk==3.11.0
pandas==2.2.2
tqdm==4.66.1
